Arreglo para evitar legajos y dni's negativos

+++ flowa/database/migrations/2024_01_11_234653_create_profesors_table.php
@@ -16,10 +16,10 @@ return new class extends Migration
             $table->timestamps();
 
             $table->string('nombre_profesor');
-            $table->bigInteger('legajo')->unique();
+            $table->unsignedInteger('legajo')->unique();
             $table->string('nombre');
             $table->string('apellido');
-            $table->bigInteger('DNI')->unique();
+            $table->unsignedInteger('DNI')->unique();
             $table->string('email')->unique();
 
         });
